feat: обновление sql скрипта.

+++ core-service/src/main/resources/db/migration/V1__init.sql
@@ -13,7 +13,7 @@ CREATE TABLE IF NOT EXISTS items(
      itle       varchar,
      image_path   varchar,
      description varchar,
-     price double,
+     price float8,
      count       int
 );
